third_party/WebKit/LayoutTests/inspector/console/console-assert-expected.txt

CONSOLE ERROR: line 9: 1
CONSOLE ERROR: line 10: a
Tests that console.assert() will dump a message and stack trace with source URLs and line numbers.

console-assert.html:9 Assertion failed: 1console-assert.html:9 bconsole-assert.html:15 a
console-assert.html:10 Assertion failed: a bconsole-assert.html:10 bconsole-assert.html:15 a
